+++ docs/tagged/cf-extension-api-v2.1.0.js
@@ -2611,4 +2611,4 @@ return /******/ (function(modules) { // webpackBootstrap
 /******/ ])
 });
 ;
-//# sourceMappingURL=cf-extension-api-v2.js.map
+//# sourceMappingURL=cf-extension-api-v2.1.0.js.map